+++ tool/shortcut.rkt
@@ -56,17 +56,27 @@
                    (send editor set-position start end)))
                (send-command "cut-clipboard" editor event))))
 ;;; c+b jump to definition
+(define jump-stack '())
 (cmd/ctrl+ "b"
            (λ (editor event)
+             (set! jump-stack (cons (send editor get-start-position) jump-stack))
              (send-command "Jump to Definition (in Other File)" editor event)
              (jump-to-definition editor (send editor get-start-position))))
 (cmd/ctrl+ "leftbutton"
            (λ (editor event)
+             (set! jump-stack (cons (send editor get-start-position) jump-stack))
              (send-command "Jump to Definition (in Other File)" editor event)
              (jump-to-definition editor
                                  (send editor find-position
                                        (send event get-x)
                                        (send event get-y)))))
+(cmd/ctrl+ "s:b"
+           (λ (editor event)
+             (if (empty? jump-stack)
+                 (void)
+                 (match-let ([(cons pop rest) jump-stack])
+                   (set! jump-stack rest)
+                   (send editor set-position pop)))))
 
 ;;; delete whole thing from current position to the start of line
 (cmd/ctrl+ "backspace"